Added Nuclei and Gf
A big update! I've replaced several mediocre manual functions with Nuclei, that can detect common issues and CVEs quickly and efficiently. By default, the scanner looks for low-high issues, but this can be changed. Also added is Gf (with Gf-Templates), which is basically an easy-to-use wrapper for Grep. Gf helps identify interesting endpoints in the (potentially huge) list of parameterized endpoints from GoCrawler and the Wayback Machine. Finally, the script now updates Go, and the TCP Nmap scan is optimized (now resolves IPS and takes only the unique ones to avoid duplicates) for a big speed improvement.