Added installer and Dockerfile
Another big change! This update adds a Dockerfile (which will be built
on Docker Hub) with all dependencies pre-installed, and a separate installer script which now includes all dependencies as well. Script configuration has been moved to a separate .env file, and the readme has been adjusted accordingly. Knowing myself this release may be buggy, so stay tuned for hotfixes :)